+++ src/main/java/com/mycompany/inventariostorepi/inventario/IInventarios.java
@@ -5,7 +5,7 @@ public interface IInventarios {
     
     public boolean agregarProducto (Inventario inventario);
     public void imprimirProductos ();
-    public void buscarProductoCodigo(double buscarCodigo);
+    public void buscarProductoCodigo(int buscarCodigo);
     public boolean editarProducto(int i, Inventario inventario );
     public boolean eliminarProducto (int i, Inventario inventario );
     public void calcularInventarioTotal ();
